pyproject.toml
[project]
name = "agentbox"
version = "0.1.0"
description = "A computer for your agent - sandboxed code execution environment"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "agentbox contributors" }
]
keywords = ["sandbox", "container", "ai", "agent", "code-execution", "gvisor", "docker"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
]
dependencies = [
    "docker>=7.0.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
]

[project.urls]
Homepage = "https://github.com/Michaelliv/agentbox"
Documentation = "https://github.com/Michaelliv/agentbox#readme"
Repository = "https://github.com/Michaelliv/agentbox"
Issues = "https://github.com/Michaelliv/agentbox/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
